+++ components/panics/include/memfault/panics/platform/reboot_tracking.h
@@ -1,0 +1,16 @@
+#pragma once
+
+//! @file
+//!
+//! Copyright (c) Memfault, Inc.
+//! See License.txt for details
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+void memfault_platform_reboot_tracking_boot(void);
+
+#ifdef __cplusplus
+}
+#endif